Hello

+++ 15/php_code.php
@@ -8,8 +8,10 @@ $time = "";
 $licence = "";
 $distance = "";
 $elapsed = "";
+$speed = "";
 $id = 0;
 $update = false;
+
 
 if (isset($_POST['save'])) {
     $date = $_POST['date'];
@@ -17,23 +19,39 @@ if (isset($_POST['save'])) {
     $licence = $_POST['licence'];
     $distance = $_POST['distance'];
     $elapsed = $_POST['elapsed'];
+    $speed = $_POST['speed'];
 
-    mysqli_query($db, "INSERT INTO info (date, time, licence, distance, elapsed) VALUES ('$date', '$time', '$licence', '$distance', '$elapsed')");
-    $_SESSION['message'] = "information saved";
+    mysqli_query($db, "INSERT INTO info (date, time, licence, distance, elapsed, speed) VALUES ('$date', '$time', '$licence', '$distance', '$elapsed', '$speed')");
+    $_SESSION['message'] = "Information Saved!";
     header('location: index.php');
 }
 if (isset($_POST['update'])) {
     $id = $_POST['id'];
-    $name = $_POST['name'];
-    $address = $_POST['address'];
+    $date = $_POST['date'];
+    $time = $_POST['time'];
+    $licence = $_POST['licence'];
+    $distance = $_POST['distance'];
+    $elapsed = $_POST['elapsed'];
+    $speed = $_POST['speed'];
 
-    mysqli_query($db, "UPDATE info SET date='$date', time='$time', licence='$licence', distance='$distance', elapsed='$elapsed' WHERE id=$id");
-    $_SESSION['message'] = "Information updated!";
+    mysqli_query($db, "UPDATE info SET date='$date', time='$time', licence='$licence', distance='$distance', elapsed='$elapsed', speed='$speed' WHERE id=$id");
+    $_SESSION['message'] = "Information Updated!";
     header('location: index.php');
 }
 if (isset($_GET['del'])) {
     $id = $_GET['del'];
+    $record = mysqli_query($db, "SELECT * FROM info WHERE id=$id");
+    $n = mysqli_fetch_array($record);
+    if (empty($n)) {
+        die('Data Not Found');
+    }
     mysqli_query($db, "DELETE FROM info WHERE id=$id");
-    $_SESSION['message'] = "Information deleted!";
+    $_SESSION['message'] = "Information Deleted!";
     header('location: index.php');
+}
+if (isset($_GET['query'])) {
+    $query = $db->escape_string($_GET['query']);
+    $sql="SELECT date, time, licence, distance, elapsed, speed FROM info WHERE licence like '%{$query}%' or date like '%{$query}%' or time like '%{$query}%' or licence like '%{$query}%' or distance like '%{$query}%' or elapsed like '%{$query}%' or speed like '%{$query}%'";
+    $results = $db->query($sql);
+    var_dump($query);
 }
